move TeleCashCurrency to own class for reuse
